face-service/requirements.txt
flask
numpy
opencv-python-headless
Pillow
requests
# Local fallback dependencies (when PYNQ is unavailable)
torch
facenet-pytorch
onnxruntime
